+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * In `CollectionsPlume`:
   * New method `isModifiable(Collection)`.
   * New methods `addIf()` and `addIfNot()`.
+  * New methods `mapListRemoveNull(Function,Iterable)` and `mapListRemoveNull(Function,Object[])`.
 * In `EntryReader`:
   * Replaced three arguments to constructors (`entryStartRegex`,
     `entryStopRegex`, `twoBlankLines`) by a single `EntryFormat` value.  The old